refactor: Rename json-converter
