+++ backend/chatsky_ui/services/json_converter/logic_component_converter/condition_converter.py
