+++ backend/chatsky_ui/services/json_converter/slots_converter.py
