+++ backend/chatsky_ui/tests/unit/test_flow_converter.py
@@ -4,10 +4,10 @@ from pathlib import Path
 import pytest
 import yaml
 
-from chatsky_ui.services.json_converter_new2.flow_converter import FlowConverter
-from chatsky_ui.services.json_converter_new2.interface_converter import InterfaceConverter
-from chatsky_ui.services.json_converter_new2.pipeline_converter import PipelineConverter
-from chatsky_ui.services.json_converter_new2.script_converter import ScriptConverter
+from chatsky_ui.services.json_converter.flow_converter import FlowConverter
+from chatsky_ui.services.json_converter.interface_converter import InterfaceConverter
+from chatsky_ui.services.json_converter.pipeline_converter import PipelineConverter
+from chatsky_ui.services.json_converter.script_converter import ScriptConverter
 
 
 @pytest.fixture
